MyClimateModel.jl API

MyClimateModel.Mesh — Type

function Mesh(nx=128)

Constructor for the mesh struct

MyClimateModel.Model — Method

function Model(mesh,...)

Constructor for the model struct

MyClimateModel.compute_equilibrium! — Function

compute_equilibrium!(max_years=100, rel_error=2e-5)

Function to run an EBM simulation until an equilibrium temperature is found

max_years is the maximum number of annual cycles to be computed when searching for equilibrium

rel_error is the tolerance for global temperature equilibrium (default is 2e-5).

MyClimateModel.compute_matrix — Method

compute_matrix(mesh,num_steps_year,model)

Function to compute the system matrix of the EBM operator

MyClimateModel.update_rhs! — Method

update_rhs!(rhs, ...)

Function to compute/update right-hand side of the system of equations that needs to be solved